requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
matplotlib>=3.5.0
pandas>=1.5.0
pandas-datareader>=0.10.0
scikit-learn>=1.1.0
yfinance>=0.2.0
MetaTrader5>=5.0.0
python-dotenv>=1.0.0
